Use `UIAlertControllerStyleAlert` on iPad to prevent crash
This crash was reported via App Store Connect and seems to
happen because the `sourceView` and `sourceRect` of the
popover view controller is not properly set.

I tried setting them up properly, but wasn't able to in a
short amount of time. So for an MVP, I decided to just use
alerts instead of an action sheet for now.