Fix the negation of `KeyValueQuery`
When an object does not have a tag with the given key,
but the query is negated, the query should match.

This is used for queries where objects that do not have the
given key-value combination are supposed to match.

In the Overpass Query Wizard, the query looks like this:
> access != private